requirements.txt
fastapi
uvicorn
pydantic
python-multipart

# parsing
pypdf
python-docx
openpyxl
python-pptx

# images
pillow

# embeddings + vector DB
faiss-cpu
numpy

# optional (better embeddings)
openai